backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# LangGraph & AI (your existing dependencies)
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Composio (your existing email integration)
composio-langchain==0.6.8
composio-core==0.6.8

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.2.0

# HTTP Client
httpx==0.27.2

# Logging
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # for TestClient

# Development
ruff==0.7.4  # Linter and formatter